+++ app/routes/show/$slug.tsx
@@ -1,35 +1,45 @@
+/* eslint-disable no-eval */
+
 import { Container, Select, Stack } from "@mantine/core";
-import type { LinksFunction, LoaderFunction } from "@remix-run/node";
+import type { LinksFunction, LoaderArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import { useFetcher, useLoaderData } from "@remix-run/react";
 import plyrStyles from "plyr/dist/plyr.css";
 import React from "react";
 import { VideoPlayer } from "~/components/VideoPlayer";
-import { getShow } from "~/helpers/get-show.helper";
-import type { EpisodeDetails } from "~/types";
+import { SHOW_DETAILS_URL } from "~/constants";
+
+type EpisodeDetails = {
+  id_episode: number;
+  season: string;
+  episode: string;
+};
 
 export const links: LinksFunction = () => {
   return [{ rel: "stylesheet", href: plyrStyles }];
 };
 
-type LoaderData = {
-  episodes: EpisodeDetails[];
-};
+export const loader = async ({ params }: LoaderArgs) => {
+  const res = await fetch(SHOW_DETAILS_URL + params.slug);
+  const markup = await res.text();
 
-export const loader: LoaderFunction = async ({ params }) => {
-  const show = await getShow(params.slug as string);
+  const SHOW_ID_REGEX = /id_show:\s(\d*),/g;
+  const iDCaptureGroup = SHOW_ID_REGEX.exec(markup)?.[1];
+  const id = iDCaptureGroup ? Number(iDCaptureGroup) : undefined;
 
-  if (!show) {
+  const SEASONS_REGEX = /seasons:\s(\[[^\]]*\])/g;
+  const seasonsCaptureGroup = SEASONS_REGEX.exec(markup)?.[1];
+  const episodes: EpisodeDetails[] = seasonsCaptureGroup ? eval(seasonsCaptureGroup) : undefined;
+
+  if (!id || !episodes) {
     throw new Response("Not Found", { status: 404 });
   }
 
-  return json<LoaderData>(show);
+  return json(episodes);
 };
 
 export default function Search() {
-  const { episodes } = useLoaderData<LoaderData>();
-
-  /** Episodes */
+  const episodes = useLoaderData<typeof loader>();
 
   const episodeOptions = episodes.map((episode) => ({
     value: episode.id_episode.toString(),
